Split large files + add native auth support for certs/serviceaccounts
