+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = easykube
-version = 0.1.0
 description = Simple, asynchronous Python client for Kubernetes.
+long_description = file: README.md
 author = Matt Pryor
 author_email = matt@stackhpc.com
 url = https://github.com/stackhpc/easykube
@@ -13,3 +13,4 @@ include_package_data = True
 packages = find:
 install_requires =
     httpx
+    pyyaml
